Change the way nodes are registered
The registration of nodes is now delayed. This is to allow the engine to know which are the enabled modules (gems) of the project before to register the nodes extending that gem (ex: EMotionFX based nodes will be enabled only when the EMotionFX gem will be enabled in the project).